removed bitDubai from fermat-dap-plugin-digital-identity-asset-user and refactor package from 'com.bitdubai' to 'org.fermat'
