+++ DAP/plugin/identity/fermat-dap-plugin-identity-asset-user/src/main/java/org/fermat/fermat_dap_plugin/layer/identity/asset/user/developer/bitdubai/version_1/database/AssetUserIdentityDatabaseConstants.java
@@ -1,0 +1,23 @@
+package org.fermat.fermat_dap_plugin.layer.identity.asset.user.developer.bitdubai.version_1.database;
+
+/**
+ * Created by Nerio on 18/09/15.
+ */
+public class AssetUserIdentityDatabaseConstants {
+    /**
+     * Asset User Identity database definition.
+     */
+    public static final String ASSET_USER_IDENTITY_DB_NAME                            = "AssetUserIdentity";
+    /**
+     * Asset User Identity database table definition.
+     */
+    public static final String ASSET_USER_IDENTITY_TABLE_NAME                         = "Asset_User_Identity";
+
+    public static final String ASSET_USER_IDENTITY_PUBLIC_KEY_COLUMN_NAME             = "public_key";
+    public static final String ASSET_USER_IDENTITY_DEVICE_USER_PUBLIC_KEY_COLUMN_NAME = "device_user_public_key";
+    public static final String ASSET_USER_IDENTITY_ALIAS_COLUMN_NAME                  = "alias";
+
+    public static final String ASSET_USER_IDENTITY_FIRST_KEY_COLUMN                   = "public_key";
+    //public static final String ASSET_USER_IDENTITY_STATUS_COLUMN_NAME = "Asset_Issuer_Identity_status";
+    //public static final String ASSET_USER_IDENTITY_PRIVATE_KEY_COLUMN_NAME = "Asset_Issuer_Identity_private_key";
+}
