+++ DAP/plugin/identity/fermat-dap-plugin-identity-asset-user/src/main/java/org/fermat/fermat_dap_plugin/layer/identity/asset/user/developer/bitdubai/version_1/exceptions/CantSetAssetUserIdentityProfileImagenException.java
@@ -1,0 +1,14 @@
+package org.fermat.fermat_dap_plugin.layer.identity.asset.user.developer.bitdubai.version_1.exceptions;
+
+import com.bitdubai.fermat_api.FermatException;
+
+/**
+ * Created by franklin on 02/11/15.
+ */
+public class CantSetAssetUserIdentityProfileImagenException extends FermatException {
+
+
+    public CantSetAssetUserIdentityProfileImagenException(final String message, final Exception cause, final String context, final String possibleReason) {
+        super(message, cause, context, possibleReason);
+    }
+}
